Remove race condition between stop() and \n in frames
When a stop() came in, it would set the terminator to \n. If one of the
pending frames had a \n in it, an exception would be thrown and it would
end up in state 3. This fixes the EOS to only be set in the state thread,
so it should always complete.